Calculadora Terminada

+++ public/javascript/calculadora.js
@@ -17,7 +17,7 @@ let funcionPredeterminada = () =>{
         let desarrolloHoras= document.getElementById('desarrolloHoras').value;
 
 
-        let resultado = parseInt(valorPorHora)* (parseInt(interfazHoras) + parseInt(testingHoras) + parseInt(estilizacionHoras) + parseInt(desarrolloHoras));
+        let resultado = parseFloat(valorPorHora)* (parseFloat(interfazHoras) + parseFloat(testingHoras) + parseFloat(estilizacionHoras) + parseInt(desarrolloHoras));
         resultado = resultado +"$";
         document.getElementById("valorTotal").value = resultado;
